// src/app/models/interfaces/user.ts
export interface User extends BaseTrackEntity {
  id: string;
  name: string;
  title: string;
  avatar: string | null;
  phone: string | null;
  email: string | null;
  description: string | null;
  companyId: string | null;
  company: Company | null;
  createdCards: TrialCaseCard[];
  deletedCards: TrialCaseCard[];
  createdCardsNotes: TrialCaseCardNote[];
  deletedCardsNotes: TrialCaseCardNote[];
  cardTasks: TrialCaseTask[];
  deletedCaseNarrativeStories: TrialCaseNarrativeStory[];
  createdTimeLineEventNotes: TimelineEventNote[];
  deletedTimeLineEventNotes: TimelineEventNote[];
  createdCases: TrialCase[] | null;
  deletedCases: TrialCase[];
  companies: Company[];
  createdFiles: File[];
  deletedFiles: File[];
  applicationUserActivities: UserActivity[];
  comments: UserComment[];
  createdTimelineEvents: TimelineEvent[];
  deletedTimelineEvents: TimelineEvent[];
  createdCaseNarrativeStories: TrialCaseNarrativeStory[];
  assignedCardTasks: TrialCaseTask[];
  modifiedCases: TrialCase[];
  modifiedCompanies: Company[];
  deletedCompanies: Company[];
  modifiedFiles: File[];
  modifiedTimelineEvents: TimelineEvent[];
  modifiedTimeLineEventNotes: TimelineEventNote[];
  modifiedCards: TrialCaseCard[];
  modifiedCardsNotes: TrialCaseCardNote[];
  createdTrialCaseFolders: TrialCaseFolder[];
  modifiedTrialCaseFolders: TrialCaseFolder[];
  deletedTrialCaseFolders: TrialCaseFolder[];
  createdTrialCasePersons: TrialCasePerson[];
  modifiedTrialCasePersons: TrialCasePerson[];
  deletedTrialCasePersons: TrialCasePerson[];
  createdTrialCaseSections: TrialCaseSection[];
  modifiedTrialCaseSections: TrialCaseSection[];
  deletedTrialCaseSections: TrialCaseSection[];
  createdTrialCaseSubsections: TrialCaseSubsection[];
  modifiedTrialCaseSubsections: TrialCaseSubsection[];
  deletedTrialCaseSubsections: TrialCaseSubsection[];
  modifiedUsers: User[];
  deletedUsers: User[];
  userTrialCaseFavorites: UserTrialCaseFavorite[];
  createdFileDocumentFolders: DocumentFolder[];
  modifiedCardTasks: TrialCaseTask[];
  deletedCardTasks: TrialCaseTask[];
  modifiedCaseNarrativeStories: TrialCaseNarrativeStory[];
  createdUsers: User[];
  modifiedFileDocumentFolders: DocumentFolder[];
  deletedFileDocumentFolders: DocumentFolder[];
  createdTrialCaseNotebookSections: TrialCaseNotebookSection[];
  modifiedTrialCaseNotebookSections: TrialCaseNotebookSection[];
  deletedTrialCaseNotebookSections: TrialCaseNotebookSection[];
  createdFileDocuments: Document[];
  modifiedFileDocuments: Document[];
  deletedFileDocuments: Document[];
  createdTrialCaseNarrativeStoryItems: TrialCaseNarrativeStoryItem[];
  modifiedTrialCaseNarrativeStoryItems: TrialCaseNarrativeStoryItem[];
  deletedTrialCaseNarrativeStoryItems: TrialCaseNarrativeStoryItem[];
  trialCaseUserRoles: TrialCaseUserRole[];
  trialCaseFolderUserRoles: TrialCaseFolderUserRole[];
}